fix: sidebar behavior
