+++ frontend/src/components/Sidebar/Sidebar.module.scss
@@ -2,7 +2,7 @@
 
 .container {
   background-color: colors.$gray-900;
-
+  position: fixed;
   height: 100vh;
   height: 100dvh;
   width: 17.75rem;
