+++ frontend/src/styles/variables/breakpoints.module.scss
@@ -1,0 +1,2 @@
+$small: 700px;
+$medium: 1200px;
